server/requirements.txt
fastapi==0.121.0
newspaper3k==0.2.8
pandas==2.3.3
Pillow==12.0.0
pydantic==2.12.3
pytesseract==0.3.13
requests==2.32.5
syllapy==0.7.2
textblob==0.19.0
uvicorn==0.38.0
lxml_html_clean
python-multipart
